+++ packages/components/forms/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.2.0](https://github.com/contentful/forma-36/compare/@contentful/f36-forms@4.1.4...@contentful/f36-forms@4.2.0) (2022-02-07)
+
+
+### Features
+
+* implement short description in page content header ([#1822](https://github.com/contentful/forma-36/issues/1822)) ([c812815](https://github.com/contentful/forma-36/commit/c812815cf6681a021fd5313fc905227def842ee8))
+
+
+
+
+
 ## [4.1.4](https://github.com/contentful/forma-36/compare/@contentful/f36-forms@4.1.3...@contentful/f36-forms@4.1.4) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-forms