+++ packages/components/icons/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.4](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.3...@contentful/f36-icons@4.0.4) (2022-02-07)
+
+
+### Bug Fixes
+
+* add preview off icon ([#1866](https://github.com/contentful/forma-36/issues/1866)) ([d88a518](https://github.com/contentful/forma-36/commit/d88a518a44e5e1a10fe3a695e4e0824e472632dc))
+
+
+
+
+
 ## [4.0.3](https://github.com/contentful/forma-36/compare/@contentful/f36-icons@4.0.2...@contentful/f36-icons@4.0.3) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-icons