+++ packages/components/note/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.7](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.6...@contentful/f36-note@4.2.7) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.6](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.5...@contentful/f36-note@4.2.6) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.5](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.4...@contentful/f36-note@4.2.5) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
+## [4.2.4](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.3...@contentful/f36-note@4.2.4) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-note
+
+
+
+
+
 ## [4.2.3](https://github.com/contentful/forma-36/compare/@contentful/f36-note@4.2.2...@contentful/f36-note@4.2.3) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-note