+++ packages/components/pill/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.8](https://github.com/contentful/forma-36/compare/@contentful/f36-pill@4.1.7...@contentful/f36-pill@4.1.8) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-pill
+
+
+
+
+
+## [4.1.7](https://github.com/contentful/forma-36/compare/@contentful/f36-pill@4.1.6...@contentful/f36-pill@4.1.7) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-pill
+
+
+
+
+
+## [4.1.6](https://github.com/contentful/forma-36/compare/@contentful/f36-pill@4.1.5...@contentful/f36-pill@4.1.6) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-pill
+
+
+
+
+
+## [4.1.5](https://github.com/contentful/forma-36/compare/@contentful/f36-pill@4.1.4...@contentful/f36-pill@4.1.5) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-pill
+
+
+
+
+
 ## [4.1.4](https://github.com/contentful/forma-36/compare/@contentful/f36-pill@4.1.3...@contentful/f36-pill@4.1.4) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-pill