+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.23](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.22...@contentful/f36-components@4.0.23) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.22](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.21...@contentful/f36-components@4.0.22) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.21](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.20...@contentful/f36-components@4.0.21) (2022-02-07)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.20](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.19...@contentful/f36-components@4.0.20) (2022-02-04)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.19](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.18...@contentful/f36-components@4.0.19) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
+## [4.0.18](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.17...@contentful/f36-components@4.0.18) (2022-02-02)
+
+**Note:** Version bump only for package @contentful/f36-components
+
+
+
+
+
 ## [4.0.17](https://github.com/contentful/forma-36/compare/@contentful/f36-components@4.0.16...@contentful/f36-components@4.0.17) (2022-01-31)
 
 **Note:** Version bump only for package @contentful/f36-components